ci: add automated release workflow and update to v1.0.0
This commit introduces automated multi-platform builds via GitHub Actions for
Linux, macOS, and Windows, and marks the first stable release.

Changes:
- Add release.yml workflow for building and publishing binaries
- Update version to 1.0.0 across Makefile, headers, and documentation
- Increase safety limits (2GB audio size, 10M ms duration)
- Update CHANGELOG with release notes
- Ignore compiled binary in gitignore

+++ wav_io.c
@@ -205,9 +205,9 @@ int load_wav_file(audio_fader_context_t *ctx, wav_header_t *header,
                   data_chunk_size, ctx->options.input_filename);
         goto cleanup;
     }
-    if (data_chunk_size > MAX_AUDIO_SIZE) {
-        LOG_ERROR("Error: Audio data size (%d bytes) exceeds maximum allowed (%d bytes)\n",
-                  data_chunk_size, MAX_AUDIO_SIZE);
+    if ((long long)data_chunk_size > MAX_AUDIO_SIZE) {
+        LOG_ERROR("Error: Audio data size (%d bytes) exceeds maximum allowed (%lld bytes)\n",
+                  data_chunk_size, (long long)MAX_AUDIO_SIZE);
         LOG_ERROR("       File: %s\n", ctx->options.input_filename);
         goto cleanup;
     }
